requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
babel==2.18.0
bcrypt==5.0.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.129.0
fastapi-cache2==0.2.2
greenlet==3.3.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
librt==0.8.1
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
mypy==1.19.1
mypy_extensions==1.1.0
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pendulum==3.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
redis==7.2.0
rsa==4.9.1
ruff==0.15.2
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
starlette-admin==0.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
